Migrated the Gurobi-Based Optimizer to a Cloud-Native application format.
Use DynamoDB as a state storage and broker. Created an "instances pump" that gets the CSV file of instances and pumps into a local and cloud DynamoDB setup, and modified the docker-compose accordingly. Included a container for a local deployment of DynamoDB.

Use DynamoDB as a semaphore/distributed lock. Deprecated filelock.py and included dynamodb_lock.py instead. Added a dockerfile to the Gurobi profile to install boto3 into Gurobi containers, to use AWS resources.

Refactored gurobi_models.py to include more metadata and all values of all variables. Refactored gurobi_optimizer.py to get the entire list of unprocessed instances from DynamoDB and try to acquire the lock of each. If fail, try the next. If success, download the instance and process it with each of the two ILP models. Store the metadata in dynamo (updating the queue) and the large data in S3.